git: 0902取得新版的 oscpu，测试没问题后进行 axi 的设计

+++ code/oscpu/libraries/difftest/src/test/csrc/verilator/emu.cpp
@@ -28,6 +28,9 @@
 #include "zlib.h"
 #include "compress.h"
 #include <list>
+#include "remote_bitbang.h"
+
+extern remote_bitbang_t * jtag;
 
 static inline void print_help(const char *file) {
   printf("Usage: %s [OPTION...]\n", file);
@@ -44,8 +47,10 @@ static inline void print_help(const char *file) {
   printf("      --load-snapshot=PATH   load snapshot from PATH\n");
   printf("      --no-snapshot          disable saving snapshots\n");
   printf("      --dump-wave            dump waveform when log is enabled\n");
+  printf("      --enable-fork          enable folking child processes to debug\n");
   printf("      --no-diff              disable differential testing\n");
   printf("      --diff=PATH            set the path of REF for differential testing\n");
+  printf("      --enable-jtag          enable remote bitbang server\n");
   printf("  -h, --help                 print program help info\n");
   printf("\n");
 }
@@ -61,6 +66,8 @@ inline EmuArgs parse_args(int argc, const char *argv[]) {
     { "force-dump-result", 0, NULL,  0  },
     { "diff",              1, NULL,  0  },
     { "no-diff",           0, NULL,  0  },
+    { "enable-fork",       0, NULL,  0  },
+    { "enable-jtag",       0, NULL,  0  },
     { "seed",              1, NULL, 's' },
     { "max-cycles",        1, NULL, 'C' },
     { "max-instr",         1, NULL, 'I' },
@@ -85,6 +92,8 @@ inline EmuArgs parse_args(int argc, const char *argv[]) {
           case 3: args.force_dump_result = true; continue;
           case 4: difftest_ref_so = optarg; continue;
           case 5: args.enable_diff = false; continue;
+          case 6: args.enable_fork = true; continue;
+          case 7: args.enable_jtag = true; continue;
         }
         // fall through
       default:
@@ -106,8 +115,6 @@ inline EmuArgs parse_args(int argc, const char *argv[]) {
     }
   }
 
-  printf("EMU, dump:%d, log_begin:%ld, log_end:%ld\n", args.enable_waveform, args.log_begin, args.log_end);
-
   Verilated::commandArgs(argc, argv); // Prepare extra args for TLMonitor
   return args;
 }
@@ -115,7 +122,7 @@ inline EmuArgs parse_args(int argc, const char *argv[]) {
 
 Emulator::Emulator(int argc, const char *argv[]):
   dut_ptr(new VSimTop),
-  cycles(0), trapCode(STATE_RUNNING), main_ticks(0)
+  cycles(0), trapCode(STATE_RUNNING)
 {
   args = parse_args(argc, argv);
 
@@ -125,36 +132,25 @@ Emulator::Emulator(int argc, const char *argv[]):
   Verilated::randReset(2);
   assert_init();
 
+  // init remote-bitbang
+  if (args.enable_jtag) { 
+    jtag = new remote_bitbang_t(23334);
+  }
   // init core
   reset_ncycles(10);
 
   // init ram
   init_ram(args.image);
 
-  difftest_init();
-  init_device();
-  if (args.enable_diff) {
-    init_goldenmem();
-    init_nemuproxy();
-  }
-
 #if VM_TRACE == 1
-#ifndef EN_FORKWAIT
-  enable_waveform = args.enable_waveform;
-  if (enable_waveform) {
+  enable_waveform = args.enable_waveform && !args.enable_fork;
+  if (enable_waveform ) {
     Verilated::traceEverOn(true);	// Verilator must compute traced signals
     tfp = new VerilatedVcdC;
     dut_ptr->trace(tfp, 99);	// Trace 99 levels of hierarchy
     time_t now = time(NULL);
     tfp->open(waveform_filename(now));	// Open the dump file
   }
-#else
-  // VM_TRACE =1 && EN_FORKWAIT
-  enable_waveform = false;
-#endif
-#else
-  // VM_TRACE =0
-  enable_waveform = false;
 #endif
 
 #ifdef VM_SAVABLE
@@ -186,27 +182,14 @@ Emulator::~Emulator() {
 }
 
 inline void Emulator::reset_ncycles(size_t cycles) {
-  dut_ptr->reset = 1;
   for(int i = 0; i < cycles; i++) {
+    dut_ptr->reset = 1;
     dut_ptr->clock = 0;
     dut_ptr->eval();
     dut_ptr->clock = 1;
     dut_ptr->eval();
-  }
     dut_ptr->reset = 0;
   }
-
-inline bool Emulator::need_dump(uint64_t * cycle_ptr) {
-  if (enable_waveform) {
-    auto trap = difftest[0]->get_trap_event();
-    uint64_t cycle = trap->cycleCnt;
-    uint64_t begin = dut_ptr->io_logCtrl_log_begin;
-    uint64_t end   = dut_ptr->io_logCtrl_log_end;
-    bool in_range  = (begin <= cycle) && (cycle <= end);
-    *cycle_ptr = cycle;
-    return in_range;
-  }
-  return false;
 }
 
 inline void Emulator::single_cycle() {
@@ -221,17 +204,19 @@ inline void Emulator::single_cycle() {
   dramsim3_helper_rising(axi);
 #endif
 
-#if VM_TRACE == 1
-  if (enable_waveform) {
-    main_ticks++;
-    if (main_ticks > args.log_begin && main_ticks < args.log_end) {
-      tfp->dump(main_ticks);
-    }
-  }
-#endif
-
   dut_ptr->clock = 1;
   dut_ptr->eval();
+
+#if VM_TRACE == 1
+  if (enable_waveform) {
+    auto trap = difftest[0]->get_trap_event();
+    uint64_t cycle = trap->cycleCnt;
+    uint64_t begin = dut_ptr->io_logCtrl_log_begin;
+    uint64_t end   = dut_ptr->io_logCtrl_log_end;
+    bool in_range  = (begin <= cycle) && (cycle <= end);
+    if (in_range) { tfp->dump(cycle); }
+  }
+#endif
 
 #ifdef WITH_DRAMSIM3
   axi_copy_from_dut_ptr(dut_ptr, axi);
@@ -241,15 +226,17 @@ inline void Emulator::single_cycle() {
   axi_set_dut_ptr(dut_ptr, axi);
 #endif
 
-#if VM_TRACE == 1
+// by ZhengpuShi, 我认为上升下降都应该捕获波形
+#if 0// VM_TRACE == 1
   if (enable_waveform) {
-    main_ticks++;
-    if (main_ticks > args.log_begin && main_ticks < args.log_end) {
-      tfp->dump(main_ticks);
-    }
+    auto trap = difftest[0]->get_trap_event();
+    uint64_t cycle = trap->cycleCnt;
+    uint64_t begin = dut_ptr->io_logCtrl_log_begin;
+    uint64_t end   = dut_ptr->io_logCtrl_log_end;
+    bool in_range  = (begin <= cycle) && (cycle <= end);
+    if (in_range) { tfp->dump(cycle); }
   }
 #endif
-
   if (dut_ptr->io_uart_out_valid) {
     printf("%c", dut_ptr->io_uart_out_ch);
     fflush(stdout);
@@ -258,13 +245,18 @@ inline void Emulator::single_cycle() {
     extern uint8_t uart_getc();
     dut_ptr->io_uart_in_ch = uart_getc();
   }
-
-  cycles++;
+  cycles ++;
 }
 
 uint64_t Emulator::execute(uint64_t max_cycle, uint64_t max_instr) {
 
-  uint32_t t = uptime();
+  difftest_init();
+  init_device();
+  if (args.enable_diff) {
+    init_goldenmem();
+    init_nemuproxy();
+  }
+
   uint32_t lasttime_poll = 0;
   uint32_t lasttime_snapshot = 0;
   uint64_t core_max_instr[NUM_CORES];
@@ -272,33 +264,37 @@ uint64_t Emulator::execute(uint64_t max_cycle, uint64_t max_instr) {
     core_max_instr[i] = max_instr;
   }
 
-#ifdef EN_FORKWAIT
-  printf("[INFO] enable fork wait..\n");
+  uint32_t t = uptime();
+  if (t - lasttime_poll > 100) {
+    poll_event();
+    lasttime_poll = t;
+  }
+
+  if(args.enable_fork){
+#ifndef EMU_THREAD
+      printf("[ERROR] please enable --threads option in verilator...(You may forget EMU_THREADS when compiling.)\n");
+      FAIT_EXIT
+#endif
+
+#ifndef VM_TRACE
+      printf("[ERROR] please enable --trace option in verilator...(You may forget EMU_TRACE when compiling.)\n");
+      FAIT_EXIT
+#endif
+
+#if EMU_THREAD <= 1
+      printf("[ERROR] please use more than 1 threads in EMU_THREADS option\n");
+      FAIT_EXIT
+#endif 
+    printf("[INFO] enable fork debugging...\n");
+  }
+
   pid_t pid =-1;
   pid_t originPID = getpid();
   int status = -1;
-  int slotCnt = 1;
+  int slotCnt = 0;
   int waitProcess = 0;
   uint32_t timer = 0;
   std::list<pid_t> pidSlot = {};
-  enable_waveform = false;
-
-  //first process as a control process
-  if ((pid = fork()) < 0) {
-    perror("First fork failed..\n");
-    FAIT_EXIT;
-  } else if (pid > 0) {  //parent process
-    printf("[%d] Control process first fork...child: %d\n ", getpid(), pid);
-    prctl(PR_SET_CHILD_SUBREAPER, 1, 0, 0, 0);
-    forkshm.shwait();
-    printf("[%d] Simulation finished, Control process exit..", getpid());
-    return cycles;
-  } else {
-    forkshm.info->exitNum++;
-    forkshm.info->flag = true;
-    pidSlot.insert(pidSlot.begin(), getpid());
-  }
-#endif
 
 #if VM_COVERAGE == 1
   // we dump coverage into files at the end
@@ -308,14 +304,12 @@ uint64_t Emulator::execute(uint64_t max_cycle, uint64_t max_instr) {
 #endif
 
   while (!Verilated::gotFinish() && trapCode == STATE_RUNNING) {
-    t = uptime();
-
-    if (t - lasttime_poll > 100) {
-      poll_event();
-      lasttime_poll = t;
+    // cycle limitation
+    if(waitProcess && cycles != 0 && cycles == forkshm.info->endCycles){
+      trapCode = STATE_ABORT;
+      break;    
     }
 
-    // cycle limitation
     if (!max_cycle) {
       trapCode = STATE_LIMIT_EXCEEDED;
       break;
@@ -341,7 +335,6 @@ uint64_t Emulator::execute(uint64_t max_cycle, uint64_t max_instr) {
     if (trapCode != STATE_RUNNING) {
       break;
     }
-
     for (int i = 0; i < NUM_CORES; i++) {
       auto trap = difftest[i]->get_trap_event();
       if (trap->instrCnt >= args.warmup_instr) {
@@ -362,7 +355,7 @@ uint64_t Emulator::execute(uint64_t max_cycle, uint64_t max_instr) {
     dut_ptr->io_perfInfo_clean = 0;
     dut_ptr->io_perfInfo_dump = 0;
 
-    if (args.enable_diff) {
+    if (args.enable_diff && !waitProcess) {
       trapCode = difftest_state();
       if (trapCode != STATE_RUNNING) break;
       if (difftest_step()) {
@@ -374,6 +367,7 @@ uint64_t Emulator::execute(uint64_t max_cycle, uint64_t max_instr) {
 #ifdef VM_SAVABLE
     static int snapshot_count = 0;
     if (args.enable_snapshot && trapCode != STATE_GOODTRAP && t - lasttime_snapshot > 1000 * SNAPSHOT_INTERVAL) {
+      // save snapshot every 60s
       time_t now = time(NULL);
       snapshot_save(snapshot_filename(now));
       lasttime_snapshot = t;
@@ -386,39 +380,47 @@ uint64_t Emulator::execute(uint64_t max_cycle, uint64_t max_instr) {
     }
 #endif
 
-#ifdef EN_FORKWAIT
-    timer = uptime();
-    if (timer - lasttime_snapshot > 1000 * FORK_INTERVAL && !waitProcess) {   // time out need to fork
-      lasttime_snapshot = timer;
-      if (slotCnt == SLOT_SIZE) {     // kill first wait process
-        pid_t temp = pidSlot.back();
-        pidSlot.pop_back();
-        kill(temp, SIGKILL);
-        slotCnt--;
-        forkshm.info->exitNum--;
-      }
-      // fork-wait
-      if ((pid = fork()) < 0) {
-        eprintf("[%d]Error: could not fork process!\n",getpid());
-        return -1;
-      } else if (pid != 0) {       // father fork and wait.
-        waitProcess = 1;
-        wait(&status);
-        enable_waveform = forkshm.info->resInfo != STATE_GOODTRAP;
-        if (enable_waveform) {
-          Verilated::traceEverOn(true);	// Verilator must compute traced signals
-          tfp = new VerilatedVcdC;
-          dut_ptr->trace(tfp, 99);	// Trace 99 levels of hierarchy
-          time_t now = time(NULL);
-          tfp->open(waveform_filename(now));	// Open the dump file
+    if(args.enable_fork){
+      timer = uptime();
+      //check if it's time to fork a checkpoint process
+      if (timer - lasttime_snapshot > 1000 * FORK_INTERVAL && !waitProcess) {   // time out need to fork
+        lasttime_snapshot = timer;
+        //kill the oldest blocked checkpoint process
+        if (slotCnt == SLOT_SIZE) {   
+          pid_t temp = pidSlot.back();
+          pidSlot.pop_back();
+          kill(temp, SIGKILL);
+          slotCnt--;
         }
-      } else {        //child insert its pid
-        slotCnt++;
-        forkshm.info->exitNum++;
-        pidSlot.insert(pidSlot.begin(), getpid());
+        //fork a new checkpoint process and block it 
+        if ((pid = fork()) < 0) {
+          eprintf("[%d]Error: could not fork process!\n",getpid());
+          return -1;
+        } else if (pid != 0) {      
+          slotCnt++;
+          pidSlot.insert(pidSlot.begin(), pid);
+        } else {     
+          waitProcess = 1;
+          uint64_t startCycle = cycles;
+          forkshm.shwait();
+          //checkpoint process wakes up
+#ifdef EMU_THREAD 
+          dut_ptr->__Vm_threadPoolp = new VlThreadPool(dut_ptr->contextp(), EMU_THREAD - 1, 0);
+#endif
+          //start wave dumping
+          enable_waveform = true;
+#if VM_TRACE == 1
+          Verilated::traceEverOn(true);	
+          tfp = new VerilatedVcdC;
+          dut_ptr->trace(tfp, 99);
+          time_t now = time(NULL);
+          tfp->open(cycle_wavefile(startCycle, now));	
+#endif
+        }
       }
     }
-#endif
+
+
 }
 
 #if VM_TRACE == 1
@@ -429,12 +431,20 @@ uint64_t Emulator::execute(uint64_t max_cycle, uint64_t max_instr) {
   save_coverage(coverage_start_time);
 #endif
 
-#ifdef EN_FORKWAIT
-  if(!waitProcess) display_trapinfo();
-  else printf("[%d] checkpoint process: dump wave complete, exit.\n",getpid());
-  forkshm.info->exitNum--;
-  forkshm.info->resInfo = trapCode;
-#endif
+  if(args.enable_fork){
+    if(waitProcess) {
+      printf("[%d] checkpoint process: dump wave complete, exit...\n",getpid());
+      return cycles;
+    }
+    else if(trapCode != STATE_GOODTRAP){
+      forkshm.info->flag = true;
+      forkshm.info->notgood = true;
+      forkshm.info->endCycles = cycles;
+      waitpid(pidSlot.back(),&status,0);
+      display_trapinfo();
+      return cycles;
+    } 
+  }
 
   display_trapinfo();
 
@@ -467,6 +477,18 @@ inline char* Emulator::waveform_filename(time_t t) {
   static char buf[1024];
   char *p = timestamp_filename(t, buf);
   strcpy(p, ".vcd");
+  printf("dump wave to %s...\n", buf);
+  return buf;
+}
+
+inline char* Emulator::cycle_wavefile(uint64_t cycles, time_t t) {
+  static char buf[1024];
+  char buf_time[64];
+  strftime(buf_time, sizeof(buf_time), "%F@%T", localtime(&t));
+  char *noop_home = getenv("NOOP_HOME");
+  assert(noop_home != NULL);
+  int len = snprintf(buf, 1024, "%s/build/%s_%ld", noop_home, buf_time, cycles);
+  strcpy(buf + len, ".vcd");
   printf("dump wave to %s...\n", buf);
   return buf;
 }
@@ -533,7 +555,6 @@ void Emulator::display_trapinfo() {
   }
 }
 
-#ifdef EN_FORKWAIT
 ForkShareMemory::ForkShareMemory() {
   if ((key_n = ftok(".", 's') < 0)) {
     perror("Fail to ftok\n");
@@ -552,9 +573,9 @@ ForkShareMemory::ForkShareMemory() {
     FAIT_EXIT
   }
 
-  info->exitNum   = 0;
   info->flag      = false;
-  info->resInfo   = -1;           //STATE_RUNNING
+  info->notgood   = false;           //STATE_RUNNING
+  info->endCycles = 0;
 }
 
 ForkShareMemory::~ForkShareMemory() {
@@ -566,23 +587,17 @@ ForkShareMemory::~ForkShareMemory() {
 
 void ForkShareMemory::shwait() {
   while (true) {
-    if (info->exitNum == 0 && info->flag) {
-      break;
+    if (info->flag ) {
+      if(info->notgood) break;
+      else exit(0);
     }
     else {
       sleep(WAIT_INTERVAL);
     }
   }
 }
-#endif
 
 #ifdef VM_SAVABLE
-
-// currently only support single core snapshot
-#if NUM_CORES != 1
-  #error "unsupported multicore"
-#endif
-
 void Emulator::snapshot_save(const char *filename) {
   static int last_slot = 0;
   VerilatedSaveMem &stream = snapshot_slot[last_slot];
@@ -668,6 +683,6 @@ void Emulator::snapshot_load(const char *filename) {
     fseek(fp, sdcard_offset, SEEK_SET);
 
   // No one uses snapshot when !has_commit, isn't it?
-  has_commit = 1;
+  diff->has_commit = 1;
 }
 #endif
